requirements.txt
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==3.7.1
asyncio==3.4.3
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dnspython==2.7.0
elevenlabs==1.53.0
fastapi==0.104.1
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
mlab_amo_async==0.0.1
motor==3.3.1
multidict==6.2.0
propcache==0.3.0
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.8.0
pymongo==4.6.0
python-dotenv==1.0.1
python-multipart==0.0.6
requests==2.31.0
sniffio==1.3.1
starlette==0.27.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.23.2
websockets==15.0.1
yarl==1.18.3
